requirements.txt
google-cloud-pubsub
google-cloud-bigquery
pandas
fastapi
uvicorn
requests
